+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- ```email``` validation rule
+- ```url``` validation rule
+
 ## [v0.1.0] 2019-03-28
 ### Added
 - CHANGELOG.md to track changes between releases of the library